[Python] Refactor save() to use file from load()
## WHAT

Modified AIConfigRuntime with an attribute file_path.
- on `config.load()` `file_path` is set to the file
Modified save() to use file path if available else default to `aiconfig.json`

## WHY
The flow from loading and saving a config should be as easy as possible. This way the user can easily update their config without having the extra step of specifying a file path