[Python][Callbacks][1/n] CallbackManager impl
For context, the callback typehint is defined as:
`Callback = Callable[[CallbackEvent], Awaitable]`

In laymans terms this means an async function that takes in a `CallbackEvent` (no output types are defined here).

Some python stuff going on here. When you define an async function, its really a function that returns a coroutine. a `coroutine` conforms to type `Awaitable`