[RFC][WIP] readme: explain 2 format modes and use cases more clearly
I took a swing at updating the AIConfig docs to reconcile artifact mode with
source-controllable mode.

The premise of this change is that ipynb is unsource-controllable, which means
AIConfig format cannot be simultaneously analogous to ipynb and config.

The answer @sarmad and I discussed is that there are 2 separate
modes or use cases each backed by the format.