[dall-e parser][2/n] Do not read streaming options
I don't think that the image generation, edit, or variations support streaming for the Dall-E 2/3 APIs. I just removed this logic because we could have streaming enabled for other prompts (ex: text-based ones)

Even for image variations (dall-e 2), the examples use a BytesIO stream, but only to get the final value of it. If you look at the API we need to pass in a full file: https://platform.openai.com/docs/guides/images/using-in-memory-image-data