[4/n] Scaffolding for Model Settings via Schema
# [4/n] Scaffolding for Components via Schema

Setting up the scaffolding of the prompt components themselves here. The main concept is that the dynamic content of each prompt (input, settings, metadata) will be rendered by either:

A) a schema representing the possible params/values and their types, with default value obtained from the config
or
B) directly from the config itself

For now, I've hard-coded an example schema used by opena model parser models. I'll look into adding some more (specifically, multi-modal) in subsequent PRs and then, if things work smoothly, we can require these schemas to be defined on each ModelParser class and then have some process to provide the static schema definitions to the client (e.g. load for each registered model parser on the backend and send that to the client once on startup).

A) will happen for newly-added prompts in the editor/workbook UI, or if existing prompts have associated schemas available.
B) can happen if no schema is provided for a model (e.g. especially in read-only cases like viewing shared gradio workbook)

One thing to be aware of for A) is that the schema represents all possible values. So, we only want to insert the values into the prompt JSON in the schema when the viewer actually sets a value in the UI. That will be handled using the `useSchemaState` hook added in this PR.

Subsequent PRs will handle setting up the same structure for prompt inputs, and then implementing the proper rendering of components for each.

## Testing:
- Just rendering the schema / state for now, making sure default values are correct:
<img width="1176" alt="Screenshot 2023-12-01 at 12 59 19 PM" src="https://github.com/lastmile-ai/aiconfig/assets/5060851/f293687a-3fdd-4e7f-8025-c240c8b6cde4">