[py] 1/n PaLM Model parser match config spec
## What

Added in a map to node api naming of completion param attributes.

## Why

PaLM API has different api across python and typescript. for example, the top k attribute

Node:
- `topK`

Python:
- `top_k`

This is also seen when exporting a workbook as aiconfig

It made sense to standardize them. the config will be able to take in `topK` and convert it to `top_k`. Both will work for py. Same for topP, top_P