2/n Runtime Interface
## What

Starting point to create a batch execution interface for AIConfig. This diff creates the interface in python. We can implement the typescript equivalent on top.

Methods Implemented:
- `config.py` (AIConfigRuntime)
- - run_batch()
- `model_parser.py` (abstract model parser class)
- - run_batch()

## Usage
User will need to construct a list of params for batch, pass it into the `run_batch` method. Very similar to the `run()` method.


Discussed interface design with @saqadri. We can iterate on top of this but this should get us started.

## Test Plan

Created a config which has a parameterized dependency chain. Batch executed on a list of params

This is for issue #