[py] Batch Execution Interface
## What

Starting point to create a batch execution interface for AIConfig. This diff creates the interface in python. We can implement the typescript equivalent on top.

Methods Implemented:
- `config.py` (AIConfigRuntime)
- - run_batch()
- `model_parser.py` (abstract model parser class)
- - run_batch()

## Usage
User will need to construct a list of params for batch, pass it into the `run_batch` method. Very similar to the `run()` method.


Discussed interface design with @saqadri. We can iterate on top of this but this should get us started.

### Note:

This diff does not implement request batching. Two things would be required on top of this diff for robust batch exec.
1. Exponential Backoff inside model parser implementations at request time. For ex: [OpenAI Backoff impl](https://platform.openai.com/docs/guides/rate-limits/error-mitigation?context=tier-free#:~:text=Example%203%3A-,Manual%20backoff%20implementation,-If%20you%20don%27t)
2. Concurrent executions. This would involve some for of multithreading. Either the model parser needs to implement custom logic to send multiple requests at once, or, the for loop that executes this batch run can become multithreaded (easy).

## Test Plan

Created a config which has a parameterized dependency chain. Batch executed on a list of params

This is for issue #