[py] Batch Execution Interface
## What

Starting point to create a batch execution interface for AIConfig. This diff creates the interface in python. We can implement the typescript equivalent on top.

Methods Implemented:
- `config.py` (AIConfigRuntime)
- - run_batch() calls `run_batch` of the model parser
- `model_parser.py` (abstract model parser class)
- - run_batch()

## Usage
User will need to construct a list of params for batch, pass it into the `run_batch` method. Very similar to the `run()` method.

Check out the implementation for more details, an example is provided in the docstring

Discussed interface design with @saqadri. We can iterate on top of this but this should get us started.

### Note:

This diff does not implement concurrent request batching. Two things would be required on top of this diff for a robust batch exec.
1. Exponential Backoff inside model parser implementations at request time. For ex: [OpenAI Backoff impl](https://platform.openai.com/docs/guides/rate-limits/error-mitigation?context=tier-free#:~:text=Example%203%3A-,Manual%20backoff%20implementation,-If%20you%20don%27t)
2. Concurrent executions. This would involve some for of multithreading. Either the model parser needs to implement custom logic to send multiple requests at once, or, the for loop that executes this batch run can become multithreaded (easy).

## Test Plan

Created a config which has a parameterized dependency chain. Batch executed on a list of params

Draft PR containing test plan: [#472](https://github.com/lastmile-ai/aiconfig/pull/472)

<img width="1629" alt="Screenshot 2023-12-14 at 2 33 28 PM" src="https://github.com/lastmile-ai/aiconfig/assets/141073967/492bbf3b-e5ba-493c-94e0-f3f57475d98c">


This is for issue [#401](https://github.com/lastmile-ai/aiconfig/issues/401)