2/n Gemini Serialize()
## What

Implemented the serialize() method of the Gemini Model parser

The gemini api is a bit trickier than previous api's we've seen before. Using the OpenAI api as a comparison point, the Gemini api is a little bit more complex.

The request takes in a contents field, which can have various types. This is because the api supports
- one shot prompts
- one shot prompting but with multiple "parts"
- multi-turn conversation with multiple parts
- multi-modal one shot with multiple parts (out of scope of this diff)

Note: `part` refers to some form of data. Either a string, a proto, a blob, or mime typed data. Gemini api has done a good job in supporting a wide array of use cases.

For simplicity, I've limited the serialize() functionality to string objects or list/dict objects with strings. If the object contains something else, an error is thrown

Notes:
- Function Calling is not supported by the public release beta. This is only available in a beta not available to the general public. Incorporating this into serialize/deserialize() may require an extra step of parsing the input. Specifically, function calling takes in a param called `tools`.


## Why

Was not implemented previously, planned as a fast follow.

## Testplan

## Issue

#463