[editor] Fix Save to Properly Serialize
# [editor] Fix Save to Properly Serialize

When opening a file in the editor we load the config into an AIConfigRuntime and send that to the client. On the client, we are modifying this serialized object and sending it to the server on save. Previously, save was just JSON serializing this to a string to save directly, which resulted in some fields (e.g. CallbackManager) being saved as well. Instead, this PR updates save to properly serialize to an AIConfig and leverages the config save method to ensure proper contents are saved to disk. As a bonus, this will also result in implicit validation of the config structure (when we add validation via the schema that will be explicit as well)

## Testing:
- Make a change in the editor & save, ensuring only that change is persisted (no more CallbackManager)