[editor] Initial Output Rendering
# [editor] Initial Output Rendering

Get prompt outputs to render. The main consideration here is that we can't really render directly based on the raw AIConfig outputs since those can be any arbitrary data defined by the model parsers. Instead, we add Client-specific types (mainly `ClientExecuteResult`) for the AIConfig which are constructed on load of the config, and which we will be able to properly handle rendering. For non-text outputs, we'll probably have an 'attachments' type to render those output attachments based on mimetype.

Side note -- maybe we just have a single type which can support both text & attachments? We should revisit if we have such models. This should do for now as a starting point.

<img width="2002" alt="Screenshot 2023-12-17 at 5 36 56 PM" src="https://github.com/lastmile-ai/aiconfig/assets/5060851/72d38c2f-59f9-4a63-8f0f-8683aecab75c">