[WIP] Add attachments field to ExecutableResult
This will mirror what we have for promptInput, and make the rendering more uniform across input and output types.


I need to change the way dalle + the huggingFaceTExt2ImageDiffusors return images. Also need to sync with Ankush to identify what the other non-text outputs are and convert those. Created issue for that in (todo):

As followups, we should add data validation.