1/n Gemini Update deserialize to take data in the form of completion params
## What

Updated deserialize() in Gemini Model Parser to take in completion data from prompt input if it is already in the form of the Gemini Completion data

Now, you can specify input in the form of a request to gemini api, ie the parts.

## Why

Previously This was skipped due to a lack of understanding of how the api requests worked.

## Testplan