1/n Gemini Update deserialize to take data in the form of completion params
## What

Updated deserialize() in Gemini Model Parser to take in completion data from prompt input if it is already in the form of the Gemini Completion data

Now, you can specify input in the form of a request to gemini api, ie the parts.

## Why

Previously This was skipped due to a lack of understanding of how the api requests worked.

## Testplan
Validating 2 pieces of functionality with this test plan:
1. setting the generation config for Gemini in setings. Generation Config is gemini's way of defining settings like topP
![testplan](https://github.com/lastmile-ai/aiconfig/assets/141073967/b120e3c8-ea1a-4c69-888e-9309d889a4f6)

2. Deserialize() can take in data just like the api, by setting "contents" in the prompt imput.