4/n Update output serialization with pb
## What

When saving outputs to config, get the underlying protobuf of Gemini's response object and save as a dict

## Why

When doing `aiconfig.save()`, pydantic is unable to deserialze the output metadata which contains a protobuf from Gemini's response. Gemini response object is a wrapper around a protobuf. Need to convert to an object that is compatible with pydantic.