[py] Batch Execution fix inconsistent outputs
## What

When Batch Executing, grab the right Prompt object to run

Also added an extra step of clearing outputs before run, similarly to #490

## Why

There was an inconsistency with the prompt object inside the deep copy and the one used to run. This diff fixes that issue.

AIConfig python sdk checks if the Prompt objects that are passed in are inside the AIConfigRuntime. It doesn't just validate name etc.

Inside batch execution, we create a deep copy of the AIConfigRuntime to run a prompt. This also means that the prompt objects are now new objects, can not be using a stale prompt object


## Testplan:

rebased @2timesJay's test diff containing an ipynb, ontop of this commit https://github.com/lastmile-ai/aiconfig/pull/553/files

I ran the ipynb with the example batch and it works as intended.

<img width="1595" alt="Screenshot 2023-12-21 at 12 15 20 PM" src="https://github.com/lastmile-ai/aiconfig/assets/141073967/0ad060f5-8019-4c61-8923-984adfb749fd">

Also compared the timings of running in batch vs in sequence. It doesn't look like batch was signficantly slower than sequence execution minus differences in gpt-outputs

<img width="1600" alt="Screenshot 2023-12-21 at 12 15 35 PM" src="https://github.com/lastmile-ai/aiconfig/assets/141073967/b01ce5cf-a72e-4f97-8789-690076b848d8">