[extensions][py][hf] 3/n Local Inference model parser changes
## What


This diff is a copy paste of the diff below except for one difference:
1. ModelParser id name and class name

HuggingFace Extension contains a `local_inference` and a `remote_inference_client`.

Remote_inference_client and the core huggingface model parser in the sdk are the same implementation with the one difference of model parser id. The extension model parser id is `HuggingFaceTextGenerationClient`.

Note: Both of these model parsers are for Text Generation only. Other huggingface tasks are not supported on the inference-endpoint by aiconfig yet.

## Why

Some context onto why there's two extensions, one was supposed to serve as a guide and easy to write model parser. This guide has yet to be written.

## Testplan

<img width="1396" alt="M" src="https://github.com/lastmile-ai/aiconfig/assets/141073967/5638c958-aa3f-4ad4-af1c-d4e733ec5722">