[editor] Functionality to run server on multiple configs without specifying port
## What:

Before creating the server process, check if the port to be used is taken. If so, find the next available one before continuing


## Why

When running the editor, user should be able to open multiple configs.

Under the hood this means spinning up another server in an unused port. Frontend already supports dynamic ports.

## Testplan

1. yarn build in aiconfig/python/src/aiconfig/editor/client
2. Open two aiconfigs with "prod" editor"

| Getting Started | Chain Of Verification |
|------------|-------------|
| ![Screenshot 1](https://github.com/lastmile-ai/aiconfig/assets/141073967/d30c9366-49d7-4b6d-851c-e469d47c4063) |  ![Screenshot 2](https://github.com/lastmile-ai/aiconfig/assets/141073967/74c53dcc-52b6-4b67-84e1-7c8900045d4a)|
| ![Screenshot 3](https://github.com/lastmile-ai/aiconfig/assets/141073967/6e86aa4f-8d7d-4140-b3ad-b374a90ae276) | ![Screenshot 4](https://github.com/lastmile-ai/aiconfig/assets/141073967/204120fa-305d-44b0-be3c-de33c342aa3c)|