Export OutputData type
Need to export this so we can set it as explicit typing and not break typescript. Also this just makes it easier to manage now.

I also rebased on top of https://github.com/lastmile-ai/aiconfig/pull/631 (changed `unknown` --> `undefined`) because the schema wasn't be generated properly otherwise

## Test plan
```
cd typescript
yarn gen-schema
```