Export OutputData type
Need to export this so we can set it as explicit typing and not break typescript. Also this just makes it easier to manage now. Also removed the `[]` form from the function_call, because that's just a value not an array.

I also rebased on top of https://github.com/lastmile-ai/aiconfig/pull/631 (changed `unknown` --> `undefined`) because the schema wasn't be generated properly otherwise

Also made the python version `OutputData` also encompass function calls just like in typescript

## Test plan
```
cd typescript
yarn gen-schema
```