Better typing for AIConfig schema
We had `JSONValue` defined as possibly `unknown` type, which made the schema generator ignore that type entirely.

This change removes that requirement, while still allowing flexibility for what gets assigned to `JSONObject`.

Regenerated the schema with `yarn gen-schema` and we get better typings now.

Note: for the `any` typing added to JSONObject, please see this workbook for what the issue is and how it's not really possible to resolve it any other way: https://lastmileai.dev/workbooks/clqnz6ebq00ekpew37w20ctzg