[editor] Connect UPDATE_PROMPT_INPUT to update_prompt endpoint
# [editor] Connect UPDATE_PROMPT_INPUT to update_prompt endpoint

Hook up the `UPDATE_PROMPT_INPUT` action to the server `update_prompt` endpoint, sending requests with a 250ms debounce so that we don't spam the server and avoid potential overlap.

We don't have server-side endpoints for updating specific parts of the prompt right now (not sure if that would be any major improvement), so we currently need to send a Prompt constructed with the updated input as a spread into the current prompt in state. We still maintain the UI state in the reducer instead of passing in this new spread prompt to the reducer from this callback since the reducer should be the source of truth for UI state and will ensure the proper immutable state flow is maintained. We can also add other state values (e.g. 'isExecuting', 'isDirty', 'isReadonly') in this reducer but leaving that out for now until it's for sure needed.

## Testing:
- Make some quick changes to the prompt input and see the proper updates and debounced requests

https://github.com/lastmile-ai/aiconfig/assets/5060851/b95d5116-9484-4d2f-88c3-de441507a140