[easy] Reference local version of aiconfig package for local editor
While we iterate on the aiconfig schema and SDK, it's better to reference the latest code for the local editor (typescript side) to ensure we are in sync on the schema updates coming in.

It also means we don't need to publish to npm/pypi until we have let the changes bake.