[editor] Prompt Name Component
# [editor] Prompt Name Component

Implement the Prompt Name component, which is currently just updating state -- need endpoint to be implemented server side before we can call it.

As part of this change, I've update the `ClientAIConfig` type to remove the `renderData` and custom output typing (in preparation for updated AIConfig output types). We still need `ClientAIConfig` and `ClientPrompt` type since our client state needs to have some custom properties. In this case, we need unique ids associated with each prompt so that we can properly render them in a list. Using name would cause the entire prompt container to fully remount every time the name input changes (and that also loses focus on the input). I've updated to use an `_ui` key on prompt for this and using helper methods for converting to/from AIConfig and ClientAIConfig (similarly for Prompt/ClientPrompt) at the boundaries.

## Testing:
- Change prompt name
- Change prompt input
- Add new prompt & change input
- Save and see the proper request is sent without _ui fields

Note: Save failing since not supplying path, will need to follow up on that


https://github.com/lastmile-ai/aiconfig/assets/5060851/f80272a5-6d74-4be2-b4cd-c5453a934c11