[editor] Save without Requiring File Path from Client
# [editor] Save without Requiring File Path from Client

The client currently doesn't know the file path that its config is saved at, so there's no way for the client to specify the 'path' to save the config at when sending the `save` request.

For now, updating the `save` endpoint to default to saving the config to the path specified in its associated AIConfigRuntime.

As we update the UX we should consider:
- Supporting changing config name/path from the client
- Sending the current path to the client on load

## Testing:
- Made changes to input (which are reflected in the server-side config)
- Click save, and ensure the config contents are saved to the config on disk