Created `get_parameters` API
We already had `get_prompt_parameters`, but this isn't good enough if parameters aren't defined locally and we want to bubble up to the aiconfig defaults.

This is important so that we can ensure we have valid prompts if not defined.

The only callsite we have today for this is in python in the `params.py` function, so I updated that callsite too. I don't know why it doesn't exist in Typescript, cc @rholingsworth

##test plan

| With local params only | With AIConfig params only | With both local and global (should use local override) | Without AIConfig but local is `{}` |
|---|---|---|---|
|||||