[python] Created `get_parameters` API
**Disclaimer:** I know this is 400+ lines changed, but ~300 of those are just adding/refactoring automated tests!!!

We already had `get_prompt_parameters`, but this isn't good enough if parameters aren't defined locally and we want to bubble up to the aiconfig defaults. See comment in https://github.com/lastmile-ai/aiconfig/pull/661#discussion_r1437941729.

This is important so that we can ensure we have valid params if not defined in the correct section of our AIConfig schema. This is similar in theme to https://github.com/lastmile-ai/aiconfig/pull/600

The only callsite we have today for this is in python in the `params.py` function, so I updated that callsite too. I don't know why it doesn't exist in Typescript (cc @rholinshead) but I also added this functionality in typescript in https://github.com/lastmile-ai/aiconfig/pull/669 (I'm not as familiar with automated testing there so will ask Ryan for a code pointer to help get me started)

## Test plan
Added a bunch of automated tests which you can run by doing going to the `aiconfig/python` dir and running `pytest`