[editor] JSON Renderer for Prompt Input
# [editor] JSON Renderer for Prompt Input

Add a toggle-able JSON editor for prompt inputs so that users can see/modify the input JSON directly. This sets up the JSONEditor component which can be reused in other places (e.g. prompt model settings). This component takes in a schema object to provide validation of the input -- we should investigate and update the implementation to just take a reference to a definition in our config schema instead of needing to duplicate the schema definition for PromptInput but I wasn't able to get that working within a reasonable time box.


https://github.com/lastmile-ai/aiconfig/assets/5060851/c5e14ac0-0771-41f5-adea-4369b886ef00