[AIC-py][editor] file path initialization
1. path is non-optional and has a default value
2. on init, create a new instance and save to the path if the file doesn't exist
(otherwise, try to load as usual)

Run server with and without explicit path. Observe in the logs
that, in either case, when the file doesn't exist it
gets created and written to.

Run again pointing at an existing valid path (e.g. the one just created).
Observe again in the logs that it loads from the path.

https://github.com/lastmile-ai/aiconfig/assets/148090348/ee24a0c0-84d8-47aa-b37f-bc9180948b4e