[editor] Connect Prompt Name to /update_prompt
# [editor] Connect Prompt Name to /update_prompt

Connect the onChangePromptName callback to the server with /update_prompt request. Note the special handling here:
Since most AIConfig API methods depend on unique prompt names, we can't update the prompt name in the config on every change (or else we could end up using a not-yet-existing prompt name for other actions). Instead, we maintain the prompt name input in the client side prompt name component state and then update the client AIConfig state after the response has returned successfully.

Note: Realized as part of this that the debounce value returns the prev debounce result, so we actually need to handle the async response dispatch event with a callback to the debounce function.

## Testing:
Make sure updating prompt name works now

https://github.com/lastmile-ai/aiconfig/assets/5060851/55da5642-9b36-44f4-aab1-0c461efab7a8