[editor] Autosave at 15s Interval
# [editor] Autosave at 15s Interval

Set up an interval to autosave the config every 15s from the client. Note that the interval is technically cleared & recreated each time the `isDirty` state changes, but that's fine because this will also handle the case where changes are made right when a successful save response comes back, or if a save fails (doesn't clear dirty state) another attempt will be made automatically after 15s.

https://github.com/lastmile-ai/aiconfig/assets/5060851/dfcb57bd-282d-4622-bd10-ec4a281c213c