[bug] Ensure that Output is constructed consistently across streaming and non-streaming
Currently the openai modelparser calls `build_output_data` to properly construct the response from the inference endpoint into an AIConfig OutputData type.

However, it only does this when stream = False. This change adds the same logic to stream = True (and also ensures that the metadata is set consistently and in a loss-less way.

Also updated the same logic for Anyscale Endpoint model parser.