Zip image response lists together for HuggingFace text to image diffusor
Victor added a helper function to aggregate the images and has_nsfw_response of each, so we don't have to do this all nested inside of a single big for-loop. Makes this easier to read and debug :)

This is mainly based on Victor's diff in https://github.com/lastmile-ai/aiconfig/pull/750, but added:

- type hints
- created an extra step where we merge the image + nsfw data into a single `ImageData` class
- [bug-fix] Moved helper function outside of class so that it can be called (if we wanted to keep it within the class, we would've had to pass `self` arg as first param and call `self.refine_responses(...)`

## Test Plan
For testing rebase on https://github.com/lastmile-ai/aiconfig/pull/461. Tested this on both StableDiffusion and StableDiffussionXL and both work

| StableDiffusion | StableDiffusionXL |
|---|---|
| <img width="1512" alt="Screenshot 2024-01-08 at 06 13 47" src="https://github.com/lastmile-ai/aiconfig/assets/151060367/f6527a48-1aa9-4c77-be56-056fcbee8ac7"> | <img width="1512" alt="Screenshot 2024-01-08 at 05 45 10" src="https://github.com/lastmile-ai/aiconfig/assets/151060367/c37647fe-e7ab-4563-b118-0e7c2bda6f79"> |