[AIC-py] cli rage
This just tells the user how to open an issue. I don't want to do anything automatically
because (a) user trust, and (b) it's much harder to implement without additional bugs.

This is mostly bedside manner. If the user is raging, they are angry.
Troll them a little bit to lighten the mood first.