[HF][fix] Remove "model" from completion params
Getting an error saying that "model" param is defined twice, which it is because we're passing it through completion params and again through the explicit model param. This along with the change in  https://github.com/lastmile-ai/aiconfig/pull/879 will ensure that the right model of `None` is passed if the value is empty string.

I also fixed typo in the aiconfig.json file