[editor][client] Enable Telemetry based on User Config settings
This diff builds on top of @rholinshead's starting point for telemetry diff. Most of that looked good to me so I didn't touch them.


Building off of #, if `allow_usage_data_sharing` is set to True, initialize Datadog Browser logging with a session ID.


Moved the initialization logic away from`index.tsx` into `editor.tsx` so that it can be configurable.

## Testplan: