Connect local HF model parser prompt input schemas to the right classnames
We did a bit of changing of classnames, so just updating here


Also important to note is that we currently don't have this for the remote inference text generation class:

https://github.com/lastmile-ai/aiconfig/blob/2364ee32d340a105b4485fa134086f9093d42384/extensions/HuggingFace/python/src/aiconfig_extension_hugging_face/remote_inference_client/text_generation.py#L124

I'll handle that next diff, as well as rename that class (and test that it works without model name specified). For now I just copy pasted the remote_inference/text_generation.py --> hf.py (core parser). Just to show that they're indeed the same functionality and that core parser can be deleted

## Test Plan
Before


After