Center the Add Prompt button
This took so long, literally this was the "how do you center a div" meme. There were basically 2 main parts that took me a while:

1. Had to update `display` from `"inline-block"` to `"flex"`. After this you don't even need to set `width: "100%"`, but I just kept it there anyways to make it obvious.
2. Even after doing that, the menu itself wasn't spread to 100%, and would only be wrapped around the `+` icon. The `Menu` component in mantine does not take regular styles (https://github.com/mantinedev/mantine/blob/703bf447b1a751daee7f3c2fff9c418bd13f5779/packages/%40mantine/core/src/components/Menu/Menu.tsx#L23-L29C24), so it took me a while to realize I had to change the child component within `Menu.Target` to `width: "100%"`

I also changed the position to be bottom centered instead of bottom left, which looks nicer imo now that this is centered.

## Test Plan
All the functionality still works

Before

After
I noticed that the GlobalParametersContainer doesn't match the same width padding/margins as the PromptContainer. This isn't usually noticeable full screen, but you can tell when you adjust the window width