2/n: Read-only mode: Connect context to Run Prompt button
Created a new value that gets read from props in `AIConfigEditor` and passed into AIConfigContext to be read by downstream components. Going to do this diff stack each component 1 at a time for easier review +  unblocking

I decided not to use a tooltip for read-only mode. Instead planning to add a big text at top saying that you're in read-only mode, and you won't be able to make any changes

## Test Plan
Still works as before

Read-only shows no tooltip

Not in read-only and running a prompt shows a tooltip