Minor nits to cancellation error message
Put the prompt name in quotes. Also specifying that we're resetting outputs only, not the entire prompt or AIConfig. Also refreshed the AIConfig to make sure that changes are propagated.

I also noticed that if you don't refresh the text lingers there. This gets fixed automatically with the next PR in diff stack!

## Test Plan

Before

After