Implement STREAM_AICONFIG_CHUNK action
This is the biggest part of this diff stack. The functionality is still exactly the same as before, but now we can call it directly instead of as a sub-action within `"CONSOLIDATE_AICONFIG"`. I think this is much cleaner.

Another benefit is this removes the need for having to check the `isRunning` flags within the consolidated `"RUN_PROMPT"` sub-action, since we now know that the only time we can call that is through the beginning and the end (non-streaming end). Next PR I am going to split this into two separate Actions:
1. `"RUN_PROMPT_START"`
2. `"RUN_PROMPT_SUCCESS"`


## Test Plan
Streaming and non-streaming models still work. Can cancel, it'll end, etc

https://github.com/lastmile-ai/aiconfig/assets/151060367/cf29ee99-9ede-4c5e-99a7-a7f7816adfe1