Fix ambiguous Image type errors
When we do
```python
from PIL import Image
```

this is the model PIL.Image. If we want the class (and type), we have to do Image.Image. I just used aliases to make things easier to read.

This wasn't erroring before because Python has fake types, and it only errored on Pydantic when we tried to create Gradio class. Probalby also tons of other runtime issues we have that Python hasn't uncovered, smh

## Test Plan
No functional changes, but it will work