[hf inference] Update HuggingFaceTextGenerationParser to support api_token in InferenceOptions
# [hf inference] Update HuggingFaceTextGenerationParser to support api_token in InferenceOptions

For Gradio, we will have user's provide an HF api token to use for running prompts and that will need to be used only for the client running the prompt for the user (so, we can't use a token from env or persist a token on the model parser instance). This PR updates the run_inference implementation of the parser to support an api_token specified in the InferenceSettings which can be used to construct the client for the text generation request.

Side note, I realize this code is a bit repetitive between the parsers but not sure the best way to reuse. This does seem like common/core functionality that could be part of the sdk itself, especially once the callbacks (e.g. for run_start) are pulled out of the model parser implementation

## Testing:
- Run text generation parser as-is, ensure it works correctly
- Update InferenceOptions in server.py with invalid 'test' api_token value and run the text generation prompt, ensuring error is propagated
<img width="1395" alt="Screenshot 2024-01-23 at 5 16 52 PM" src="https://github.com/lastmile-ai/aiconfig/assets/5060851/0e1ca23a-feba-4367-b674-d409e34a785a">
- Update InferenceOptions with valid api_token and ensure it runs correctly